chore : 날짜 포맷 유틸 함수 분리
